requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

ultralytics==8.1.0
numpy==1.26.4
pydantic==2.5.3
python-dotenv